allminds/requirements.txt
[packages]
amqp==2.5.1
Babel==2.7.0
billiard==3.6.1.0
bootstrap4==0.1.0
cachetools==3.1.1
celery==4.3.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
cssselect==1.1.0
cssutils==1.0.2
dateparser==0.7.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.6
django-appconf==1.0.3
django-auditlog==0.4.5
django-bulk-update==2.2.0
django-celery-beat==1.5.0
django-chartjs==1.5.0
django-ckeditor==5.7.1
django-filter==2.2.0
django-formtools==2.1
django-imagekit==4.0.2
-e git+https://github.com/django-import-export/django-import-export.git@3cf5e3f9796a5caf7c5ea3928119af7ce4706c0d#egg=django_import_export
django-js-asset==1.2.2
django-jsonfield==1.3.1
django-material==1.5.8
django-money==0.15.1
django-otp==0.7.3
django-phonenumber-field==1.3.0
django-request==1.5.5
django-select2==7.1.1
django-simple-mail==2.2.6
django-timezone-field==3.1
django-two-factor-auth==1.9.1
django-viewflow==1.5.6
et-xmlfile==1.0.1
html2text==2019.9.26
humanize==0.5.1
idna==2.8
importlib-metadata==0.23
jdcal==1.4.1
kombu==4.6.5
lxml==4.4.1
MarkupPy==1.14
maya==0.6.1
more-itertools==7.2.0
odfpy==1.4.0
openpyxl==3.0.0
pendulum==2.0.5
phonenumbers==8.10.22
pilkit==2.0
Pillow==6.2.1
premailer==3.6.1
psycopg2==2.8.4
py-moneyed==0.8.0
python-crontab==2.4.0
python-dateutil==2.6.0
python-decouple==3.1
pytz==2019.3
pytzdata==2019.3
PyYAML==5.1.2
qrcode==6.1
regex==2019.11.1
requests==2.22.0
six==1.12.0
snaptime==0.2.4
sqlparse==0.3.0
tablib==0.14.0
tzlocal==2.0.0
urllib3==1.25.6
vine==1.3.0
xlrd==1.2.0
xlwt==1.3.0
zipp==0.6.0